asm, vm: conditional execution
Implemented conditional execution which is read on the first 4 bytes of
the instruction. When the conditional fields are set it will check the
result of the previous S operation (e.g. cmp, subs, adds, etc) and
compare it against 0. If the condition is not met the instruction is
ignored. Conditions are appended to the mnemonic e.g. mov if equal: moveq.

Implemented new opcode:

cmp: sets the conditional value to Ds - Ops1

Implemented conidions:

gt:   greater than
lt:   less than
eq:   equal
gteq: greater than or equal
lteq: less than or equal