asm, vm: implemented new memory model
The new stack mechanism is pre allocated upon creating the VM. The
instructions push and pop allow you to push to the stack and pop from
the stack. Push and pop are pseudo op codes and are translated in to
several opcodes. The stack is a fully decrementing stack where pushing
decrements the SP (stack pointer, r13) prior to storing the new value.